LayoutTests/fast/forms/file/entries-api/image-no-transcode-open-panel-expected.txt

OPEN FILE PANEL
Tests a HEIC image, chosen via the file picker which accepts HEIC, will not be converted to any supported MIME type.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS fileList.length is 1
PASS file.type is "image/heic"
PASS file.size is non-zero.
Image could not be loaded.
PASS img.width is not 400
PASS img.height is not 400
PASS successfullyParsed is true

TEST COMPLETE
  
